Am donat mongo

+++ pages/api/products/[id].js
@@ -2,7 +2,9 @@ import dbConnect from '../../../lib/mongo'
 import Product from '../../../models/Product'
 
 export default async function handler(req, res) {
-    const {method, query: {id}} = req;
+    const {method, query: {id}, cookies} = req;
+
+    const token = cookies.token
 
     dbConnect();
 
@@ -15,6 +17,9 @@ export default async function handler(req, res) {
         }
     }
     if(method === "PUT"){
+        if(!token){
+            return res.status(401).json({message: "Not authorized"})
+        }
         try{
             const product = await Product.findByIdAndUpdate(id, res.body, {new: true});
             res.status(201).json(product);
@@ -23,6 +28,9 @@ export default async function handler(req, res) {
         }
     }
     if(method === "DELETE"){
+        if(!token){
+            return res.status(401).json({message: "Not authorized"})
+        }
         try{
             await Product.findByIdAndDelete(id);
             res.status(201).json("The product has been deleted");
